Convert the assetic adapter, RelativePosition & Dependency to traits.
