+++ core/lib/Drupal/Core/Asset/Aggregate/AggregateAsset.php
@@ -16,11 +16,13 @@ use Drupal\Core\Asset\Collection\BasicAssetCollection;
 use Drupal\Core\Asset\Exception\AssetTypeMismatchException;
 use Drupal\Core\Asset\Exception\UnsupportedAsseticBehaviorException;
 use Drupal\Core\Asset\Metadata\AssetMetadataInterface;
+use Drupal\Core\Asset\AsseticAdapterTrait;
 
 /**
  * Base class for representing aggregate assets.
  */
 class AggregateAsset extends BasicAssetCollection implements \IteratorAggregate, AssetInterface, AggregateAssetInterface {
+  use AsseticAdapterTrait;
 
   /**
    * @var \Drupal\Core\Asset\Metadata\AssetMetadataInterface
@@ -262,33 +264,4 @@ class AggregateAsset extends BasicAssetCollection implements \IteratorAggregate,
   public function clearFilters() {
     $this->filters->clear();
   }
-
-  /**
-   * @throws \Drupal\Core\Asset\Exception\UnsupportedAsseticBehaviorException
-   */
-  public final function getVars() {
-    throw new UnsupportedAsseticBehaviorException("Drupal does not use or support Assetic's 'vars' concept.");
-  }
-
-  /**
-   * @throws \Drupal\Core\Asset\Exception\UnsupportedAsseticBehaviorException
-   */
-  public final function setValues(array $values) {
-    throw new UnsupportedAsseticBehaviorException("Drupal does not use or support Assetic's 'values' concept.");
-  }
-
-  /**
-   * @throws \Drupal\Core\Asset\Exception\UnsupportedAsseticBehaviorException
-   */
-  public final function getValues() {
-    throw new UnsupportedAsseticBehaviorException("Drupal does not use or support Assetic's 'values' concept.");
-  }
-
-  /**
-   * @throws \Drupal\Core\Asset\Exception\UnsupportedAsseticBehaviorException
-   */
-  public final function getLastModified() {
-    throw new UnsupportedAsseticBehaviorException("Drupal does not use or support Assetic's getLastModified() concept.");
-  }
-
 }
