+++ core/tests/Drupal/Tests/Core/Asset/Aggregate/AggregateAssetTest.php
@@ -300,26 +300,5 @@ class AggregateAssetTest extends AssetUnitTest {
   public function testDump() {
     $this->fail();
   }
-
-  /**
-   * @expectedException \Drupal\Core\Asset\Exception\UnsupportedAsseticBehaviorException
-   */
-  public function testGetVars() {
-    $this->getAggregate()->getVars();
-  }
-
-  /**
-   * @expectedException \Drupal\Core\Asset\Exception\UnsupportedAsseticBehaviorException
-   */
-  public function testSetValues() {
-    $this->getAggregate()->setValues(array());
-  }
-
-  /**
-   * @expectedException \Drupal\Core\Asset\Exception\UnsupportedAsseticBehaviorException
-   */
-  public function testGetValues() {
-    $this->getAggregate()->getValues();
-  }
 }
